src/xOps/requirements.txt
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
Django==3.0.7
django-redis==4.12.1
djangorestframework==3.11.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.5
